requirements.txt
flask
blis==0.7.9
thinc==8.1.9
spacy==3.4.4
requests 
beautifulsoup4 
fpdf2==2.5.6 
gtts 
pydub 
spotipy 
openai
python-dotenv